Local xsmb

+++ lrvappl/app/Http/Controllers/PostsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use Illuminate\Support\Facades\DB;
 
@@ -14,6 +15,12 @@ class PostsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+    
+     public function __construct()
+    {
+        $this->middleware('auth', ['except'=>['index','show']]);
+    }
+
     public function index()
     {
         // $post = Post::all(); // get all properties of POST   
@@ -46,12 +53,28 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title' =>'required',
-            'body'=>'required'
+            'body'=>'required',
+            'cover_image'=>'image|nullable|max:1999'
         ]);
+
+        if ($request->hasFile('cover_image')) {
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $fileName =pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $fileExt = $request->file('cover_image')->getClientOriginalExtension();
+
+            $fileNameToStore = $fileName.'_'.time().'.'.$fileExt;
+
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        } else {
+            $fileNameToStore = 'no_image.jpg';
+        }
+        
 
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id =auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         return redirect('/posts')->with('success','New post created');
     }
@@ -77,6 +100,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error','Unauthorized Post');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -91,12 +117,29 @@ class PostsController extends Controller
     {
         $this->validate($request,[
             'title' =>'required',
-            'body'=>'required'
+            'body'=>'required',
+            'cover_image'=>'image|nullable|max:1999'
         ]);
 
+        if ($request->hasFile('cover_image')) {
+            $fileNameWithExt = $request->file('cover_image')->getClientOriginalName();
+            $fileName =pathinfo($fileNameWithExt, PATHINFO_FILENAME);
+            $fileExt = $request->file('cover_image')->getClientOriginalExtension();
+
+            $fileNameToStore = $fileName.'_'.time().'.'.$fileExt;
+
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+
         $post = Post::find($id);
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error','Unauthorized Post');
+        }
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         return redirect('/posts')->with('success','Post updated');
     }
@@ -110,6 +153,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (auth()->user()->id !== $post->user_id) {
+            return redirect('posts')->with('error','Unauthorized Post');
+        }
+        if($post->cover_image != 'no_image.jpg'){
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
